+++ pkg/ibmvpc/subnetsIPBlocks_test.go
@@ -4,7 +4,7 @@ Copyright 2023- IBM Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package vpcmodel
+package ibmvpc
 
 import (
 	"slices"
